+++ mayan/apps/messaging/links.py
@@ -63,7 +63,7 @@ link_message_single_delete = Link(
 link_message_list = Link(
     condition=condition_user_is_authenticated,
     badge_text=get_unread_message_count, icon=icon_message_list,
-    text='', view='messaging:message_list'
+    title=_(message='Messages'), view='messaging:message_list'
 )
 link_message_single_mark_read = Link(
     args='object.pk', conditional_disable=condition_is_read,